Move shared function to command.go for readability
